refactor: funding 통합테스트 경로 변경 # 👆 여기 위에 커밋 메시지를 작성하세요 # 주석부는 자동으로 지워집니다 # 형식: <타입>: <제목> (최대 50자) # 본문: (선택사항): 무엇을, 왜 변경했는지 설명 # Resolves #123 (선택적)
# ============================================
# 👇아래는 참고용 가이드 (자동 제거됨)
# ============================================
#
# 작성 예시:
# Feat: 로그인 API 추가
#
# JWT 기반 인증 방식으로 구현
# - Spring Security 설정
# - 토큰 생성/검증 로직
#
# Resolves #123
# ============================================

# Feat:     새로운 기능 추가
# Fix:      버그 수정
# Docs:     문서 수정
# Style:    포맷팅, 세미콜론 누락 (코드 변경 X)
# Refactor: 리팩토링
# Test:     테스트 코드
# Chore:    빌드, 패키지 매니저 수정
# Design:   UI 디자인 변경
# Comment:  주석 추가/변경
# Rename:   파일/폴더 이름 변경
# Remove:   파일 삭제
# !HOTFIX:  긴급 버그 수정
# !BREAKING CHANGE: API 변경
#
# 팁: 제목 대문자 시작, 마침표 X, 명령조
